add string normalization function
We will need to compare the word chosen by one player with the guess
provided by another player. Players might not write the same word the
same way but the words must still be considered equal. For instance, if
player1 chose "accusé" and player2 guesses "ACCUSE", it should be a
match.

In order to do so, we introduce a normalization function which transform
strings into a normalized form. We will then compare the normalized
forms instead of the original strings.